Implement callback for persistent jobs
Uses reflection to avoid dependency on persistent jobs - will attach the event handler if persistent jobs is loaded.
This might be able to be done with Harmony, I'm not sure, but I'm 99% sure that would require at least a compile time reference to persistent jobs so I didn't go down that path.

This version simply accepts a job has been changed, logs the fact, and then triggers a reload of and publish of jobs - this effectively updates the relevant tracks correctly.